// Frontend/front/src/app/api/model/paymentDto.ts
/**
 * CubArt API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PaymentStatusEnum } from './paymentStatusEnum';
import { PaymentMethodEnum } from './paymentMethodEnum';


export interface PaymentDto { 
    id: string;
    purchaseId: string;
    purchaseName: string;
    amount: number;
    paymentMethod: PaymentMethodEnum;
    paymentStatus: PaymentStatusEnum;
    dateCreated: string;
}
export namespace PaymentDto {
}